sample-content-repo/image_classifier/requirements.txt

torch
torchvision
numpy
matplotlib
pillow
